requirements.txt
transformers
streamlit
requests
python-dotenv